Fixed issues for the iOS platform.
Fixed the issue with getting the application path on iOS, the file managers
`currentDirectoryPath` did not work as expected. This behavior have been
replaced with the NSBundle `resourcePath`.

The implementation for checking which dispatch queue is currently running
have been modified since `dispatch_get_current_queue` have been removed from
iOS 7. This have been replaced with setting the name to the container for
the dispatch queue.