Modifications to `RASqlite+RATable` and added `RASqliteTableDelegate`.
The `RASqlite+RATable`-category have been renamed to `RASqlite+RASqliteTable`,
to minimize the chance of naming collisions.

The interface for the category have been modified to include the upcoming
changes for easier table checking. Additional argument have been added to the
`checkTable:withColumns:`-method to include the the table status. However, the
support for the status flag have yet to be implemented.

The `check`-method have been modified to take into account the new protocol
`RASqliteTableDelegate`, which will ease the handling of table modifications.

The idea behind the protocol is to easier enable developers to migrate or make
modifications to data or tables before/after tables have been checked, created,
or removed. As of now, only parts of the check handling have been implemented.