move bind statement to separate class
in an effort to reduce the internal complexity of the `RASqlite`-class,
the code for binding parameters to statements have moved to a separate
class. this will also improve testability of the new class.

the method signature have been slightly modified to allow `RASqlite`
access to an possible error, the internal method remains the same.